json data moved to raw-data and imported from that file instead for reus abliity
